Create linux spec for runc spec command
